% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggr_es.R
\name{aggr_es}
\alias{aggr_es}
\title{Aggregates event-study treatment effects.}
\usage{
aggr_es(
  object,
  rhs = 0,
  period = c("post", "pre", "both"),
  aggregation = c("mean", "cumulative"),
  abbr_term = TRUE,
  ...
)
}
\arguments{
\item{object}{A model object of class \code{fixest}, where the \code{i()} operator has
been used to facilitate an "event-study" DiD design. See Examples.}

\item{rhs}{Numeric. The null hypothesis value. Defaults to 0.}

\item{aggregation}{Character. The aggregation type. Either "mean" (the
default) or "cumulative".}

\item{abbr_term}{Logical. Should the leading "term" column of the return
data frame be abbreviated? The default is TRUE. If FALSE, then the term
column will retain the full hypothesis test string as per usual with
\code{marginaleffects()}. Note that this information is retained as an attribute
of the return object, regardless.}

\item{...}{Additional arguments passed to \code{marginaleffects::deltamethod()}.}
}
\value{
A "tidy" data frame of aggregated (pre and/or post) treatment
effects, plus inferential information about standard errors, confidence
intervals, etc. Potentially useful information about the underlying
hypothesis test is also provided as an attribute. See Examples.
}
\description{
Aggregates post- (and/or pre-) treatment effects of an
"event-study" estimation, also known as a dynamic difference-in-differences
(DDiD) model. The event-study should have been estimated using the \code{fixest}
package, which provides a specialised \code{i()} operator for this class
of models. By default, the function will return the average post-treatment
effect (i.e. across multiple periods). However, it can also return the
cumulative post-treatment effect and can be used to aggregate pre-treatment
effects too. At its heart, \code{aggr_es()} is a convenience wrapper around
\code{marginaleffects::deltamethod()}, which is used to perform the underlying
joint hypothesis test.
}
\examples{
library(fixest)

est = feols(y ~ x1 + i(period, treat, 5) | id + period, base_did)

# Default hypothesis test is a null mean post-treatment effect
(post_mean = aggr_es(est))
# The underlying hypothesis is saved as an attribute
attributes(post_mean)["hypothesis"]

# Other hypothesis and aggregation options
aggr_es(est, aggregation = "cumulative")
aggr_es(est, period = "both")
aggr_es(est, rhs = -1, period = "pre")
# Etc.

}
\seealso{
\code{\link[marginaleffects:deltamethod]{marginaleffects::deltamethod()}}
}
